# library-managment-api/static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Library Management API",
      "description": "API for managing a collection of books in a library.",
      "version": "1.0.0"
    },
    "host": "localhost:5000",
    "basePath": "/",
    "schemes": ["http"],
    "paths": {
      "/books": {
        "get": {
          "summary": "List all books",
          "responses": {
            "200": {
              "description": "A list of books",
              "schema": {
                "type": "array",
                "items": { "$ref": "#/definitions/Book" }
              }
            }
          }
        },
        "post": {
          "summary": "Add a new book",
          "parameters": [
            {
              "in": "body",
              "name": "book",
              "required": true,
              "schema": { "$ref": "#/definitions/Book" }
            }
          ],
          "responses": {
            "201": { "description": "Book added successfully" },
            "400": { "description": "Invalid input" }
          }
        }
      },
      "/books/search": {
        "get": {
          "summary": "Search for books",
          "parameters": [
            { "in": "query", "name": "author", "type": "string" },
            { "in": "query", "name": "published_year", "type": "string" },
            { "in": "query", "name": "genre", "type": "string" }
          ],
          "responses": {
            "200": {
              "description": "Filtered list of books",
              "schema": {
                "type": "array",
                "items": { "$ref": "#/definitions/Book" }
              }
            }
          }
        }
      },
      "/books/{isbn}": {
        "delete": {
          "summary": "Delete a book by ISBN",
          "parameters": [
            { "in": "path", "name": "isbn", "required": true, "type": "string" }
          ],
          "responses": {
            "200": { "description": "Book deleted successfully" },
            "404": { "description": "Book not found" }
          }
        },
        "put": {
          "summary": "Update a book by ISBN",
          "parameters": [
            { "in": "path", "name": "isbn", "required": true, "type": "string" },
            {
              "in": "body",
              "name": "book",
              "required": true,
              "schema": { "$ref": "#/definitions/Book" }
            }
          ],
          "responses": {
            "200": { "description": "Book updated successfully" },
            "404": { "description": "Book not found" }
          }
        }
      }
    },
    "definitions": {
      "Book": {
        "type": "object",
        "required": ["title", "author", "published_year", "isbn"],
        "properties": {
          "title": { "type": "string" },
          "author": { "type": "string" },
          "published_year": { "type": "integer" },
          "isbn": { "type": "string" },
          "genre": { "type": "string" }
        }
      }
    }
  }
  